Move LocalizationMap to api
